Move lambdas to src directory so tests run and because it makes more sense
