--- lambdas/tsconfig.json
@@ -1,6 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "include": [
-    "*.ts"
-  ]
-}
