+++ src/lambdas/decorate-pools.ts
@@ -3,8 +3,8 @@
  * This lambda runs after prices have been updated, it updates the liquidity of each pool
  * based on the latest token prices + pool token values.
  */
-import { getPools, getTokens, updatePools } from "../src/data-providers/dynamodb";
-import { PoolDecorator } from "../src/pools/pool.decorator";
+import { getPools, getTokens, updatePools } from "../data-providers/dynamodb";
+import { PoolDecorator } from "../pools/pool.decorator";
  
 export const handler = async (): Promise<any> => {
   const log = console.log;
