+++ src/lambdas/get-pool.ts
@@ -1,4 +1,4 @@
-import { getPool } from '../src/data-providers/dynamodb';
+import { getPool } from '../data-providers/dynamodb';
 
 export const handler = async (event: any = {}): Promise<any> => {
   const chainId = parseInt(event.pathParameters.chainId);
