+++ src/lambdas/get-tokens.ts
@@ -1,4 +1,4 @@
-import { getTokens } from '../src/data-providers/dynamodb';
+import { getTokens } from '../data-providers/dynamodb';
 
 export const handler = async (event: any = {}): Promise<any> => {
   const chainId = parseInt(event.pathParameters.chainId);
