+++ src/lambdas/tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "extends": "../../tsconfig.json",
+  "include": [
+    "*.ts"
+  ]
+}
