+++ src/lambdas/update-prices.ts
@@ -1,5 +1,5 @@
-import { getTokens } from "../src/data-providers/dynamodb";
-import { updateTokenPrices } from "../src/tokens";
+import { getTokens } from "../data-providers/dynamodb";
+import { updateTokenPrices } from "../tokens";
 
 export const handler = async (): Promise<any> => {
   const log = console.log;
